% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateBalance.R
\name{computeCovariateBalance}
\alias{computeCovariateBalance}
\title{Compute covariate balance for each comparison}
\usage{
computeCovariateBalance(indication = "Depression", outputFolder, maxCores)
}
\arguments{
\item{indication}{A string denoting the indication for which the exposure cohorts should be created.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}
}
\description{
Compute covariate balance for each comparison
}
\details{
Compute covariate alance for each comparison. The balance is computed considering the entire target and comparator cohort,
so note removing subjects with prior outcomes. Stores the results in a subfolder called 'balance'.
}
