% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incidence.R
\name{computeIncidence}
\alias{computeIncidence}
\title{Compute incidence}
\usage{
computeIncidence(indicationId = "Depression", outputFolder)
}
\arguments{
\item{indicationId}{A string denoting the indicationId for which the exposure cohorts should be
created.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes (/)}
}
\description{
Compute incidence
}
\details{
Compute incidence using the CohortMethod data files.
}
