% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidenceRates.R
\name{computeIncidenceRates}
\alias{computeIncidenceRates}
\title{Compute incidence rates}
\usage{
computeIncidenceRates(indication = "Depression", outputFolder)
}
\arguments{
\item{indication}{A string denoting the indication for which the exposure cohorts should be created.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}
}
\description{
Compute incidence rates
}
\details{
Compute incidence rates using the CohortMethod data files.
}
