% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadDataToDb.R
\name{createIndicesOnDatabase}
\alias{createIndicesOnDatabase}
\title{Create indices in the LEGEND result database}
\usage{
createIndicesOnDatabase(connectionDetails, staging = FALSE)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created
using the
\code{\link[DatabaseConnector]{createConnectionDetails}}
function in the DatabaseConnector package.}
}
\description{
Create indices on the large tables in the LEGEND evidence model for faster searching.
}
