% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrioExposureCovariateBuilder.R
\name{createPriorExposureCovariateSettings}
\alias{createPriorExposureCovariateSettings}
\title{Create settings for adding prior exposures as covariates}
\usage{
createPriorExposureCovariateSettings(cohortDatabaseSchema = "unknown",
  exposureEraTable = "unknown", windowStart = -999, windowEnd = -1,
  analysisId = 999)
}
\arguments{
\item{cohortDatabaseSchema}{The name of the database schema that is the location
where the data used to define the exposure cohorts is
available.}

\item{exposureEraTable}{The tablename that contains the exposure eras.}

\item{windowStart}{Start day of the window where covariates are captured,
relative to the index date (0 = index date).}

\item{windowEnd}{End day of the window where covariates are captured,
relative to the index date (0 = index date).}

\item{analysisId}{A unique identifier for this analysis.}
}
\value{
A covariateSettings object.
}
\description{
Create settings for adding prior exposures as covariates
}
\details{
Counts the number of prior treatments.
}
