% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubgroupCovariateBuilder.R
\name{createSubgroupCovariateSettings}
\alias{createSubgroupCovariateSettings}
\title{Create settings for adding subgroup markers as covariates}
\usage{
createSubgroupCovariateSettings(windowStart = -365, windowEnd = -1,
  analysisId = 998)
}
\arguments{
\item{windowStart}{Start day of the window where covariates are captured,
relative to the index date (0 = index date).}

\item{windowEnd}{End day of the window where covariates are captured,
relative to the index date (0 = index date).}

\item{analysisId}{A unique identifier for this analysis.}
}
\value{
A covariateSettings object.
}
\description{
Create settings for adding subgroup markers as covariates
}
\details{
Creates a marker covariate for each of the subgroups of interest.
}
