% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{legend_report}
\alias{legend_report}
\title{OHDSI-formatted LEGEND report}
\usage{
legend_report(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Boolean toggle as to whether intermediate LaTex file is saved}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
OHDSI-formatted LEGEND report
}
\section{Document options}{


\describe{
  \item{\code{databaseId}}{TODO}
}
}

