% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadDataToDb.R
\name{unstageDb}
\alias{unstageDb}
\title{Drop the old tables, and remove _staging postfix from staging tables.}
\usage{
unstageDb(connectionDetails, schema = "legend")
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created
using the
\code{\link[DatabaseConnector]{createConnectionDetails}}
function in the DatabaseConnector package.}

\item{schema}{The schema holding the LEGEND results.}
}
\description{
Drop the old tables, and remove _staging postfix from staging tables.
}
