% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomCmDataObjectBuilding.R
\name{generateAllCohortMethodDataObjects}
\alias{generateAllCohortMethodDataObjects}
\title{Construct all cohortMethodData object}
\usage{
generateAllCohortMethodDataObjects(outputFolder, indicationId = "Depression")
}
\arguments{
\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}

\item{indicationId}{A string denoting the indicationId.}
}
\description{
Construct all cohortMethodData object
}
\details{
This function constructs all cohortMethodData objects using the data
fetched earlier using the \code{\link{fetchAllDataFromServer}} function.
}
