% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadDataToDb.R
\name{uploadResultsToDatabase}
\alias{uploadResultsToDatabase}
\title{Upload results to the database server}
\usage{
uploadResultsToDatabase(connectionDetails, exportFolder,
  createTables = FALSE, staging = FALSE)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created
using the
\code{\link[DatabaseConnector]{createConnectionDetails}}
function in the DatabaseConnector package.}

\item{exportFolder}{Name of local folder where the export results are stored; make sure to use
forward slashes (/). Do not use a folder on a network
drive since this greatly impacts performance.}

\item{createTables}{Create the tables on the server? This will drop the tables
if they already exist prior to creating them.}

\item{staging}{If TRUE, the table names on the server will have the postfix "_staging" to
allow new tables to be created while old ones are in use.}
}
\description{
Upload the exported results of a database-indication run into the database. Assumes the data conforms to
the LEGEND data model.
}
